obd-ii: Add support for decoding 29bit query and responses
This patch adds support for decoding 29bit CAN IDs. Much of the
existing code was in place for handling 29bit IDs but lacked the
ability to check for the correct 29bit request and response IDs.
This patch adds that ability and correctly selects for use of either
11bit or 29bit CAN IDs.

Change-Id: I7cf10a56aa93d951c3ffa45734139689b3f3af4c
Reviewed-on: https://code.wireshark.org/review/34297
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>