rsl: Add support for non-standard RSL_MSG_TYPE_OSMO_ETWS_CMD
3GPP sadly has forgotten to specify how the ETWS Primary Notification
is to be communicated over Abis to the BTS.  Osmocom has introduced
a non-standard RSL message type "RSL_MSG_TYPE_OSMO_ETWS_CMD" for this
purpose to the OsmoBTS and OsmoBSC implementations.

It contains a single Information Element which in turn contains the
56-byte primary notification of the Earthquake and Tsunami Warning
system.

This patch adds wireshark dissector support to it.

Change-Id: Icce5b06bfd00c0a7ef0f21283cc57473fa34f034
Reviewed-on: https://code.wireshark.org/review/34464
Petri-Dish: Pascal Quantin <pascal@wireshark.org>
Tested-by: Petri Dish Buildbot
Reviewed-by: Pascal Quantin <pascal@wireshark.org>