Couchbase: Handle extras in mutation responses
Previously extras in any mutation (e.g. ADD, SET) response would
be marked as illegal, as these used to return no extras.
More recently, extras can be returned in mutations if the
connection has negotiated the MUTATION_SEQNO feature.
The extras can contain the mutation's sequence number and the
vBucket UUID.
This commit adds a new field (couchbase.extras.mutation_seqno) to
achieve this, as the existing field couchbase.extras.seqno is used
for a different purpose and is only a UINT32, whereas a UINT64 is
returned as part of the extras; the existing
couchbase.extras.vbucket_uuid is used for the vBucket UUID.

Change-Id: If8a5148f2115fce7a777b96ad22ba92d95c9ff71
Reviewed-on: https://code.wireshark.org/review/34540
Reviewed-by: Dave Rigby <daver@couchbase.com>
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>