Fix build with non-standard install prefix
The variables BUILD_dumpcap and PCAP_FOUND are defined after the if
condition, therefore wireshark is built with a relative RPATH and
library loading for privileged processes fails with glibc.

Move the definition of BUILD_dumpcap before it is used and drop
the test for PCAP_FOUND.

Move a comment to the proper line while at it.

Change-Id: Ie151a3781898e712f9e324e9ef995022bbea40b6
Reviewed-on: https://code.wireshark.org/review/34564
Petri-Dish: João Valverde <j@v6e.pt>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>
Reviewed-on: https://code.wireshark.org/review/34572
Reviewed-by: João Valverde <j@v6e.pt>