NSIS: Tell the user to uninstall Npcap 0.99[45] manually.
Npcap 0.994 and 0.995 might crash the system while being removed. Later
versions of the installer warn about this, but a MessageBox doesn't
provide much space to describe what's going on.

Add manual uninstallation instructions to the release notes.

If we find a problematic version of Npcap in the NSIS installer, tell
the user that they should uninstall it manually. Link to the relevant
section in the release notes.

Bug: 16052
Change-Id: Ie8d811b5169ef253ffbd24a5f24a470b3a846e4f
Reviewed-on: https://code.wireshark.org/review/34565
Reviewed-by: Gerald Combs <gerald@wireshark.org>
Petri-Dish: Gerald Combs <gerald@wireshark.org>
Tested-by: Petri Dish Buildbot
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>