Improve the handling of the FT (Frame Type) field in Type III headers.
Show the value symbolically, and don't dissect the payload as an
Ethernet packet if the value isn't 0, meaning "Ethernet".

This gets rid of the mis-dissection of the payload in the capture
atttached to bug 16089, although it doesn't dissect it any further, as
we don't know how to dissect it.

Change-Id: I97fce0f7a4f4336339bc90271aa7b19c97831abe
Ping-Bug: 16089
Reviewed-on: https://code.wireshark.org/review/34649
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 3367cbab17769658872decc78762c8edf681fe43)
Reviewed-on: https://code.wireshark.org/review/34662