GSMTAP: implement dissecting of PTCCH/D messages
According to 3GPP TS 45.002, section 3.3.4.2, PTCCH (Packet Timing
advance control channel) is a packet dedicated channel, that is
used for continuous Timing Advance control in (E)GPRS.

There are two sub-types of that logical channel:

  - PTCCH/U (Uplink): used to transmit random Access Bursts
    to allow estimation of the Timing Advance for one MS in
    packet transfer mode.

  - PTCCH/D (Downlink): used by the network to transmit
    Timing Advance updates for several MS.

As per 3GPP TS 45.003, section 5.2, the coding scheme used for
PTCCH/U is the same as for PRACH as specified in subclause 5.3,
while the coding scheme used for PTCCH/D is the same as for
CS-1 as specified in subclause 5.1.1.

The format of PTCCH/D messages can be described as follows:

  +--------------+--------------+-----+---------------+------------------+
  |    Octet 1   |    Octet 2   |     |    Octet 16   |  Octet 17 .. 23  |
  +---+----------+---+----------+-----+---+-----------+------------------+
  | 0 | TA TAI=0 | 0 | TA TAI=1 | ... | 0 | TA TAI=15 | Padding 00101011 |
  +---+----------+---+----------+-----+---+-----------+------------------+

what gives us 16 Timing Advance values (7 bit each) for 16 different
mobile stations identified by TAI (0..15). The remaining space is
padding and shall be filled with constant value 0x2B.

Bug: 16096
Change-Id: I0ce81d922a8a8c3981da2486baa3e1efcff46539
Reviewed-on: https://code.wireshark.org/review/34660
Reviewed-by: Pau Espin Pedrol <pespin@sysmocom.de>
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>