Fix checks for "do we have an ERSPAN header?"
For a GRE protocol type of 0x008E, check the "sequence number present"
bit in the GRE header, if it's available, to see if the packet has an
ERSPAN header or not, rather than checking the entire header to see if
it's zero.  (If the GRE header isn't available, assume no ERSPAN
header.)

For a GRE protocol type of 0x22EB, always treat the packet as having an
ERSPAN header.

That matches more closely what the most recent I-D for ERSPAN said.

Bug: 16089
Change-Id: I21119411e8485854fca85fa701b994bfa4e73941
Reviewed-on: https://code.wireshark.org/review/34664
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot
Reviewed-by: Guy Harris <guy@alum.mit.edu>