IPv6: New Hop by Hop Path MTU Option
Add support to Wireshark for new hop-by-hop option defined in:

https://tools.ietf.org/html/draft-ietf-6man-mtu-option-01

This Hop-by-Hop IPv6 option is designed to be used to
record the minimum Path MTU along the forward path between a source
host to a destination host.  This collects a minimum recorded MTU
along the path to the destination.  The value can then be
communicated back to the source using the return Path MTU field in
the option.

Change-Id: I5dfc1a70aedd347b128bc3c5cd96f9d5a0752e3b
Reviewed-on: https://code.wireshark.org/review/34786
Reviewed-by: João Valverde <j@v6e.pt>
Petri-Dish: João Valverde <j@v6e.pt>
Tested-by: Petri Dish Buildbot
Reviewed-by: Michael Tüxen <tuexen@wireshark.org>