QUIC: atag is an array of bytes, not pointers
The allocated buffer is unnecessarily large due to the array storing
pointers instead of bytes. This wastes memory but has no other bad
side-effects. It was not caught by the compiler because the users accept
void pointers. I only noticed it in the debugger.

Change-Id: I9690b5481289d17fed34512b6a32915c3a30d36b
Reviewed-on: https://code.wireshark.org/review/35184
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>