LNET: fix malformed packet error on RDMA Read Request
Check the number of remaining bytes before comparing the
next four bytes to the LNet magic number to find out
whether or not this is an LNet layer.
In the case for an RDMA Read Request which does not have
any payload data, the LNet dissector now gracefully will
not process this packet as an LNet layer instead of
labeling it as a malformed packet.

Bug: 16269
Change-Id: I5d6abea8b79619849e8fbc1bc6b585995736b56e
Reviewed-on: https://code.wireshark.org/review/35392
Reviewed-by: Alexis La Goutte <alexis.lagoutte@gmail.com>
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>