Introduce AudioCodes Debug Recording (ACDR) dissector
ACDR is a protocol over UDP that is used by AudioCodes devices for
recording traffic to and from the device.

It adds a header to each packet that contains extra data about the packet.

For some packet types (like SIP), it also appends the IP and UDP/TCP
headers of the sent/received packet.

The dissector unwraps the ACDR header, and displays the packets with the
original type (and when available, with the original addresses).

Bug: 16275
Change-Id: I19ad90053a2ef73da80881dc5e94aa362de23ea3
Reviewed-on: https://code.wireshark.org/review/35417
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot
Reviewed-by: Peter Wu <peter@lekensteyn.nl>