MPLS: fix Downstream Detailed Mapping TLV label stack dissection
The label stack is its own TLV, which adheres to the same format as
others, with a tag and length (obviously). The length to be used for
this TLV is not the remaining length of this packet, but its own length.

Bug: 16273
Change-Id: I469b55c5d78d1899d87030f3c0a15fd1afdf3f85
Reviewed-on: https://code.wireshark.org/review/35436
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Tested-by: Petri Dish Buildbot
Reviewed-by: Anders Broman <a.broman58@gmail.com>
(cherry picked from commit 6d7e98dd2b8fe577af3985611a922f9905a1a262)
Reviewed-on: https://code.wireshark.org/review/35451