afs: add some "new" RPCs
Teach the dissector about some newer RPCs, and document some of the gaps
in the RPC ranges.

While here, remove a few commented-out lines that refer to previous
incarnations of the reassembly code.

Also correct the ouput description for the Link RPC, which actually
describes the status of the created (hard) link, not a symlink.

Change-Id: I03181cb4807f842842fe8a1003fc7b40a49ad2ba
Reviewed-on: https://code.wireshark.org/review/36298
Reviewed-by: Anders Broman <a.broman58@gmail.com>