rx: decode version packets
An RX_PACKET_TYPE_VERSION packet may be used to request the AFS version
from a server.  It may also be issued periodically as a "NAT ping", to
assist in keeping UDP port mappings intact over what may otherwise be a
long period of network inactivity.

Provide some information in the info column for each of these use cases.

Change-Id: I73aa7f2cbfd4a3e9538d629086bfa35eda6b6bf7
Reviewed-on: https://code.wireshark.org/review/36294
Reviewed-by: Anders Broman <a.broman58@gmail.com>